services/image-gen/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
diffusers==0.27.2
transformers==4.39.3
accelerate==0.29.3
torch==2.2.2               # 自动安装 CPU 版
Pillow==10.3.0
numpy==1.26.4